Time: 2 ms (29.97%), Space: 23 MB (97.42%) - LeetHub

+++ 3242-count-elements-with-maximum-frequency/3242-count-elements-with-maximum-frequency.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int maxFrequencyElements(vector<int>& nums) {
-        uint8_t freq[101] = {0};
-        uint8_t max = 0, res = 0;
+        int freq[101] = {0};
+        int max = 0, res = 0;
         for (int n : nums) {
-            uint8_t f = ++freq[n];
+            int f = ++freq[n];
             if (f > max) {
                 max = f;
                 res = f;
